Adding tags support to schema.
Tags aren't supported in VIRL, yet, but we wanted to add them as
part of the upcoming release's schema change.  Note that to ensure
that the tags element would immediately follow its parent element,
I changed how extensions are declared.  The XML instance documents
should be the same, but the schema declaration now uses groups
instead of extensible types.